# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QueueLess Pulse - Smart Queue Management for Banks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /swap/list:
    post:
      operationId: createSwapListing
      tags: [swap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId]
              properties:
                bookingId: { type: integer }
                note: { type: string }
      responses:
        "200":
          description: Listed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SwapListing" }
  /swap/{listingId}/accept:
    post:
      operationId: acceptSwap
      tags: [swap]
      parameters:
        - in: path
          name: listingId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [myBookingId]
              properties:
                myBookingId: { type: integer }
      responses:
        "200":
          description: Swap completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /swap/{listingId}:
    delete:
      operationId: cancelSwapListing
      tags: [swap]
      parameters:
        - in: path
          name: listingId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    SwapListing:
      type: object
      required: [id, bookingId, ownerName, branchName, serviceName, bookingDate, timeSlot, tokenNumber, note, createdAt]
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        ownerName: { type: string }
        branchName: { type: string }
        serviceName: { type: string }
        bookingDate: { type: string }
        timeSlot: { type: string }
        tokenNumber: { type: string }
        note: { type: string }
        createdAt: { type: string }
...